Move button into /components/atoms for better hierarchy of components
