+++ app/components/pages/homepage.tsx
@@ -3,7 +3,7 @@ import Faq from '../templates/faq';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import { Grid, Box } from '@mui/material';
-import Navbar from '../header';
+import Navbar from '../templates/header';
 import BackgroundVideo from '../BackgroundVideo';
 import { Link } from '@remix-run/react';
 
@@ -95,9 +95,11 @@ export default function Homepage() {
               </h2>
             </div>
             <div className="text-center">
-              <button className="px-8 py-4 mt-8 text-xl text-white rounded-md bg-seacrest-350 font-poppins hover:bg-seacrest-100">
+              <Link
+                to="/"
+                className="px-8 py-4 mt-8 text-xl text-white rounded-md bg-seacrest-350 font-poppins hover:bg-seacrest-100">
                 Ready to start?
-              </button>
+              </Link>
             </div>
             <div className="mt-8 text-center text-white">
               Not ready? <br /> Try a{' '}
